refactor(*): 패키지 위치와 패키지명을 정리정
