+++ src/main/java/racingcar/Application.java
@@ -1,9 +1,9 @@
 package racingcar;
 
 import camp.nextstep.edu.missionutils.Console;
-import racingcar.business.CarRacingService;
-import racingcar.controller.CarRacingController;
-import racingcar.object.carEngine.RegulationsCarEngine;
+import racingcar.racingapp.business.CarRacingService;
+import racingcar.racingapp.controller.CarRacingController;
+import racingcar.racingapp.model.car.carEngine.RegulationsCarEngine;
 
 public class Application {
