+++ src/main/java/racingcar/racingapp/model/car/carEngine/RegulationsCarEngine.java
@@ -1,4 +1,4 @@
-package racingcar.object.carEngine;
+package racingcar.racingapp.model.car.carEngine;
 
 import camp.nextstep.edu.missionutils.Randoms;
