+++ src/main/java/racingcar/racingapp/model/value/enums/DriveResult.java
@@ -1,4 +1,4 @@
-package racingcar.object.enums;
+package racingcar.racingapp.model.value.enums;
 
 public enum DriveResult {
     SUCCESS,
