+++ src/main/java/racingcar/racingapp/model/value/enums/RaceMaxCounts.java
@@ -1,11 +1,11 @@
-package racingcar.object.enums;
+package racingcar.racingapp.model.value.enums;
 
 public enum RaceMaxCounts {
 
     MAX_COUNT_OF_CAR(5),
     MAX_COUNT_OF_LAP(10);
 
-    private Integer count;
+    private final Integer count;
 
     RaceMaxCounts(final Integer count) {
         this.count = count;
