+++ src/main/java/racingcar/racingapp/model/value/vo/CarRacingResult.java
@@ -1,4 +1,4 @@
-package racingcar.object.value;
+package racingcar.racingapp.model.value.vo;
 
 public record CarRacingResult(RaceRecord raceRecord, Winners winners) {
