+++ src/main/java/racingcar/racingapp/model/value/vo/CarRecord.java
@@ -1,4 +1,4 @@
-package racingcar.object.value;
+package racingcar.racingapp.model.value.vo;
 
 public record CarRecord(CarName carName, Mileage mileage) {
