+++ src/main/java/racingcar/racingapp/model/value/vo/Winners.java
@@ -1,4 +1,4 @@
-package racingcar.object.value;
+package racingcar.racingapp.model.value.vo;
 
 public record Winners(CarNames winners) {
