+++ src/test/java/racingcar/object/carEngine/BrokenCarEngine.java
@@ -1,5 +1,7 @@
 package racingcar.object.carEngine;
 
+import racingcar.racingapp.model.car.carEngine.CarEngine;
+
 public class BrokenCarEngine implements CarEngine {
 
     @Override
