+++ src/test/java/racingcar/object/value/MileageTest.java
@@ -4,7 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.Test;
-import racingcar.object.enums.RaceMaxCounts;
+import racingcar.racingapp.model.value.enums.RaceMaxCounts;
+import racingcar.racingapp.model.value.vo.Mileage;
 
 class MileageTest {
